# API-Gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Freelance Marketplace.",
        "title": "GigForge",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "gigforge.jasim.online",
    "basePath": "/",
    "paths": {
...
        "/client/projects/buy/{id}": {
            "post": {
                "security": [
                    {
                        "ClientAccessToken": []
                    }
                ],
                "description": "Buy a specific project",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Buy a project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully bought project",
                        "schema": {
                            "$ref": "#/definitions/project.BuyProjectRes"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/freelancer/projects": {
            "get": {
                "security": [
                    {
                        "FreelancerAccessToken": []
                    }
                ],
                "description": "Get a list of projects posted by the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "List my projects",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved projects",
                        "schema": {
                            "$ref": "#/definitions/project.ListMyProjectRes"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "FreelancerAccessToken": []
                    }
                ],
                "description": "Create a new project listing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Add a single project",
                "parameters": [
                    {
                        "description": "Project details",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.AddSingleProject"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully added project",
                        "schema": {
                            "$ref": "#/definitions/project.AddSingleProjectRes"
                        }
                    },
                    "400": {
                        "description": "Error parsing request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Validation errors",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/freelancer/projects/{id}": {
            "patch": {
                "security": [
                    {
                        "FreelancerAccessToken": []
                    }
                ],
                "description": "Edit an existing project listing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Edit a project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Project details",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.AddSingleProject"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully edited project",
                        "schema": {
                            "$ref": "#/definitions/project.EditSingleProjectRes"
                        }
                    },
                    "400": {
                        "description": "Error parsing request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Validation errors",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/freelancer/projects/{projectID}": {
            "delete": {
                "security": [
                    {
                        "FreelancerAccessToken": []
                    }
                ],
                "description": "Remove an existing project listing",
                "tags": [
                    "projects"
                ],
                "summary": "Remove a project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully removed project",
                        "schema": {
                            "$ref": "#/definitions/project.RemProjectRes"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/projects": {
            "get": {
                "description": "Get a list of all projects",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "List all projects",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved projects",
                        "schema": {
                            "$ref": "#/definitions/project.ListProjectsRes"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/projects/payment": {
            "get": {
                "description": "Get details of a project payment",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Get project payment details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "orderID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment project details",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/projects/payment/capture": {
            "post": {
                "description": "Capture payment for a project order",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Capture project payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "paymentID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "orderID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully captured payment",
                        "schema": {
                            "$ref": "#/definitions/project.CapturePaymentRes"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/projects/payment/execute": {
            "post": {
                "description": "Execute payment for a project order",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Execute project payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "orderID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully executed payment",
                        "schema": {
                            "$ref": "#/definitions/project.ExecutePaymentRes"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "description": "Get details of a specific project",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Get project by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved project",
                        "schema": {
                            "$ref": "#/definitions/project.ListOneProjectRes"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "project.AddSingleProjectRes": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "Response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "project.BuyProjectRes": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "Response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "project.CapturePaymentRes": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "project.EditSingleProjectRes": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "Response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "project.ExecutePaymentRes": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "PaymentID": {
                    "type": "string"
                },
                "merchantID": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "project.ListMyProjectRes": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "Project": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/project.Project"
                    }
                },
                "Response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "project.ListOneProjectRes": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "Response": {
                    "type": "string"
                },
                "project": {
                    "$ref": "#/definitions/project.Project"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "project.ListProjectsRes": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "Response": {
                    "type": "string"
                },
                "project": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/project.Project"
                    }
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "project.Project": {
            "type": "object",
            "properties": {
                "Category": {
                    "type": "string"
                },
                "Description": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "Price": {
                    "type": "number"
                },
                "Title": {
                    "type": "string"
                },
                "delivery_days": {
                    "type": "integer"
                },
                "number_of_revisions": {
                    "type": "integer"
                }
            }
        },
        "project.RemProjectRes": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "Response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "req.AddSingleProject": {
            "type": "object",
            "required": [
                "Description",
                "category",
                "title"
            ],
            "properties": {
                "Delivery days": {
                    "type": "integer"
                },
                "Description": {
                    "type": "string"
                },
                "Number_of_revisions": {
                    "type": "integer"
                },
                "category": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ClientAccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "FreelancerAccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}